Feat: Implement CardFormBrick View
