+++ Sources/MercadoPagoCheckout/Views/InstallmentView.swift
@@ -4,13 +4,15 @@
 //
 //  Created by Guilherme Prata Costa on 14/11/25.
 //
+
 import MPComponents
+import MPFoundation
 import SwiftUI
 
 struct InstallmentScreen: View {
     var body: some View {
         MPHeader(
-            title: "Product",
+            title: MPStrings.Installments.title,
             onBack: {
                 print("Back tapped")
             }
